Create workspace if it doesn't exist
Plugin failed on report download if started with new job which
has no workspace present. The Jenkins doc says that the required
node() should allocate workspace but it's not always available
so might be bug in Jenkins side or there's some missing configuration
in Job init so this might need revisit.